Update Phonebook.c++
made many changes. In the original version, for some reason, whenever I tried to add more than one name to the book, the same name would get added again. I could not find the root of the problem so I ended up restructuring the classes to try a different approach that ended up working. I commented out the original and left it in just in case. Still need to add the function to change a person's phone number and save to a text file